[리팩토링] 구조 변경
