--- BlackCat/Filter/FilterCell/FilterBaseCell.swift
@@ -1,13 +1,0 @@
-//
-//  FilterBaseCell.swift
-//  BlackCat
-//
-//  Created by Hamlit Jason on 2022/10/15.
-//
-
-import UIKit
-import RxSwift
-
-class FilterBaseCell: BaseCollectionViewCell {
-    var disposeBag: DisposeBag = DisposeBag()
-}
